voice-agent/pyproject.toml
 [build-system]
 requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"

 [project]
 name = "wordpan-voice-agent"
 version = "0.1.0"
 description = "LiveKit voice agent for WordPan vocabulary practice"
 requires-python = ">=3.9"

 dependencies = [
     "av==15.1.0",
     "livekit-agents[silero,turn-detector]~=1.2",
     "livekit-plugins-noise-cancellation~=0.2",
     "python-dotenv",
    "supabase>=2.22.0",
 ]

 [dependency-groups]
 dev = [
     "pytest",
     "pytest-asyncio",
     "ruff",
     "watchdog>=6.0.0",
 ]

 [tool.setuptools.packages.find]
 where = ["src"]

 [tool.setuptools.package-dir]
 "" = "src"

 [tool.pytest.ini_options]
 asyncio_mode = "auto"
 asyncio_default_fixture_loop_scope = "function"

 [tool.ruff]
 line-length = 88
 target-version = "py39"

 [tool.ruff.lint]
 select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
 ignore = ["E501"]  # Line too long (handled by formatter)

 [tool.ruff.format]
 quote-style = "double"
 indent-style = "space"

